requirements.txt
pip==23.2.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.11.2
setuptools==68.1.2
wheel==0.41.1
tqdm==4.66.1
